+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * New builtin: `\lastValue`.
 * A new mode `last` is added. It outputs the last possible result. You can switch to this mode by the `-t` flag in the command line or the `\Mode last` command in the REPL.
+* Now input can have leading and trailing spaces. The interpreter will trim them when parsing.
+* Now Nekomata can be built with GHC 9.10.1, thanks to an update of the `haskeline` package.
 
 ### Breaking changes